requirements.txt
autopep8==1.7.0
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg==3.1.13
psycopg-binary==3.1.13
pycodestyle==2.9.1
python-dotenv==1.0.0
toml==0.10.2
typing_extensions==4.8.0
Werkzeug==2.2.2
zipp==3.8.1
